+++ test/test_helper.exs
@@ -34,6 +34,8 @@ ExUnit.start(
     # :pix_domain,
     # :static_brcode,
     # :dynamic_brcode,
+    # :pix_key_holmes,
+    # :pix_key_holmes_log,
     # :issuing_balance,
     # :issuing_product,
     # :issuing_card,
@@ -74,6 +76,7 @@ Code.require_file("./test/utils/pix_pull_request.exs")
 Code.require_file("./test/utils/random.exs")
 Code.require_file("./test/utils/static_brcode.exs")
 Code.require_file("./test/utils/dynamic_brcode.exs")
+Code.require_file("./test/utils/pix_key_holmes.exs")
 Code.require_file("./test/utils/ledger.exs")
 Code.require_file("./test/utils/ledger_transaction.exs")
 Code.require_file("./test/utils/issuing_holder.exs")